requirements.txt
python-telegram-bot[job-queue,socks]>=20.0
pandas
openpyxl
python-dotenv
filelock
pytz
